Correction on the id argument in the ClientRepository.findOneById method: the database store id as long type

+++ app/src/main/java/com/holcvart/androidptut/model/repository/ClientRepository.java
@@ -27,7 +27,7 @@ public class ClientRepository extends EntityRepository{
         client.setId(database.insert(PhoneRepairManagementContract.Client.TABLE_NAME, null, values));
     }
 
-    public Client findOneById(int id){
+    public Client findOneById(long id){
         String selection = PhoneRepairManagementContract.Client._ID + " = ?";
         String[] selectionArgs = { String.valueOf(id) };
         Cursor cursor=database.query
